KMS-672: Refactor DIF10 corrections to use shared XML path editor
